social-media/requirements.txt
asyncpg==0.29.0
fastapi==0.110.0
Flask_RESTful==0.3.10
passlib==1.7.4
pydantic==2.6.3
SQLAlchemy==2.0.23
typer==0.9.0
uvicorn==0.28.0
grpcio==1.39.0
protobuf==3.20.0
pydantic_settings==2.2.1